更新

+++ zheng-upms/zheng-upms-server/src/main/java/com/zheng/upms/server/controller/manage/UpmsRoleController.java
@@ -85,11 +85,13 @@ public class UpmsRoleController extends BaseController {
             }
         }
         // 删除权限
-        UpmsRolePermissionExample upmsRolePermissionExample = new UpmsRolePermissionExample();
-        upmsRolePermissionExample.createCriteria()
-            .andPermissionIdIn(deleteIds)
-            .andRoleIdEqualTo(id);
-        upmsRolePermissionService.deleteByExample(upmsRolePermissionExample);
+        if (deleteIds.size() > 0) {
+            UpmsRolePermissionExample upmsRolePermissionExample = new UpmsRolePermissionExample();
+            upmsRolePermissionExample.createCriteria()
+                    .andPermissionIdIn(deleteIds)
+                    .andRoleIdEqualTo(id);
+            upmsRolePermissionService.deleteByExample(upmsRolePermissionExample);
+        }
         return new UpmsResult(UpmsResultConstant.SUCCESS, datas.size());
     }
